util/cpuset: Add Linux build constraint
CPU affinity primitives are currently only available on Linux,
viewed at cmd/vendor/golang.org/x/sys/unix/affinity_linux.go.
Other unix systems will default to no-op.
